requirements.txt
polars
python-box[all]
tqdm
